+++ provider/pkg/gen/schema.go
@@ -144,14 +144,18 @@ func PulumiSchema(swagger map[string]interface{}) pschema.PackageSpec {
 					Description: kind.Comment() + kind.PulumiComment(),
 					Type:        "object",
 					Properties:  map[string]pschema.PropertySpec{},
+					Language:    map[string]json.RawMessage{},
 				}
 
+				var propNames []string
 				for _, p := range kind.Properties() {
 					objectSpec.Properties[p.name] = genPropertySpec(p, kind.canonicalGV, kind.kind)
+					propNames = append(propNames, p.name)
 				}
 				for _, p := range kind.RequiredInputProperties() {
 					objectSpec.Required = append(objectSpec.Required, p.name)
 				}
+				objectSpec.Language["nodejs"] = rawMessage(map[string][]string{"requiredOutputs": propNames})
 
 				pkg.Types[tok] = objectSpec
 				if kind.IsNested() {
@@ -182,6 +186,9 @@ func PulumiSchema(swagger map[string]interface{}) pschema.PackageSpec {
 		}
 	}
 
+	// Compatibility mode for Kubernetes 2.0 SDK
+	const kubernetes20 = "kubernetes20"
+
 	pkg.Language["csharp"] = rawMessage(map[string]interface{}{
 		"packageReferences": map[string]string{
 			"Glob":                         "1.1.5",
@@ -189,7 +196,7 @@ func PulumiSchema(swagger map[string]interface{}) pschema.PackageSpec {
 			"System.Collections.Immutable": "1.6.0",
 		},
 		"namespaces":             csharpNamespaces,
-		"compatibility":          "kubernetes20",
+		"compatibility":          kubernetes20,
 		"dictionaryConstructors": true,
 	})
 	pkg.Language["go"] = rawMessage(map[string]interface{}{
@@ -198,6 +205,7 @@ func PulumiSchema(swagger map[string]interface{}) pschema.PackageSpec {
 		"packageImportAliases": pkgImportAliases,
 	})
 	pkg.Language["nodejs"] = rawMessage(map[string]interface{}{
+		"compatibility": kubernetes20,
 		"dependencies": map[string]string{
 			"@pulumi/pulumi":    "^2.0.0",
 			"shell-quote":       "^1.6.1",
@@ -214,6 +222,21 @@ func PulumiSchema(swagger map[string]interface{}) pschema.PackageSpec {
 			"@types/shell-quote": "^1.6.0",
 		},
 		"moduleToPackage": modToPkg,
+		"readme": `The Kubernetes provider package offers support for all Kubernetes resources and their properties.
+Resources are exposed as types from modules based on Kubernetes API groups such as 'apps', 'core',
+'rbac', and 'storage', among many others. Additionally, support for deploying Helm charts ('helm')
+and YAML files ('yaml') is available in this package. Using this package allows you to
+programmatically declare instances of any Kubernetes resources and any supported resource version
+using infrastructure as code, which Pulumi then uses to drive the Kubernetes API.
+
+If this is your first time using this package, these two resources may be helpful:
+
+* [Kubernetes Getting Started Guide](https://www.pulumi.com/docs/quickstart/kubernetes/): Get up and running quickly.
+* [Kubernetes Pulumi Setup Documentation](https://www.pulumi.com/docs/quickstart/kubernetes/configure/): How to configure Pulumi
+    for use with your Kubernetes cluster.
+
+Use the navigation below to see detailed documentation for each of the supported Kubernetes resources.
+`,
 	})
 	pkg.Language["python"] = rawMessage(map[string]interface{}{
 		"requires": map[string]string{